03 - django_lms/requirements.txt
crispy-bootstrap5==0.6
Django==3.2.9
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-extensions==3.1.5
django-filter==21.1
Faker==9.8.3
flake8==4.0.1
flake8-builtins==1.5.3
flake8-import-order==0.18.1
ipython==7.30.0
pep8-naming==0.12.1
python-dateutil==2.8.2
webargs==8.0.1
